Moved relational algebra to reasoning algorithm
